+++ app/Http/Controllers/CommentaireController.php
@@ -4,12 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Commentaire;
 use App\Article;
+use App\User;
 use Illuminate\Support\Facades\Auth;
 
 use Illuminate\Http\Request;
 
 class CommentaireController extends Controller
 {
+    // public function __construct(){
+    //     $this->middleware('isAdminOrWebmaster')->only('index',);
+    //     $this->middleware('isAdminOrWebmasterOrMe')->only('edit','update','destroy');
+    // }
     /**
      * Display a listing of the resource.
      *
@@ -17,6 +22,7 @@ class CommentaireController extends Controller
      */
     public function index()
     {
+        $this->authorize('adminWebmaser', User::class);
         $commentaires = Commentaire::all();
         return view('commentaire.index',compact('commentaires'));
     }
@@ -28,7 +34,7 @@ class CommentaireController extends Controller
      */
     public function create()
     {
-        //    
+        return redirect()->back();   
     }
 
     /**
@@ -55,6 +61,7 @@ class CommentaireController extends Controller
      */
     public function edit(Commentaire $commentaire)
     {
+        $this->authorize('adminWebWritter',$commentaire, Commentaire::class);
         return view('commentaire.edit',compact('commentaire'));
     }
 
@@ -67,6 +74,7 @@ class CommentaireController extends Controller
      */
     public function update(Request $request, Commentaire $commentaire)
     {
+        $this->authorize('adminWebWritter',$commentaire, Commentaire::class);
         $commentaire->commentaire = $request->input('commentaire');
         $commentaire->save();
         return redirect()->back()->with('edit-commentaire', 'Votre message a été modifié!');
@@ -81,6 +89,7 @@ class CommentaireController extends Controller
      */
     public function destroy(Commentaire $commentaire)
     {
+        $this->authorize('adminWebWritter',$commentaire, Commentaire::class);
         $commentaire->delete();
         return redirect()->back()->with('destroy-commentaire', 'Votre message a été supprimé!');
     }
